+++ package.json
@@ -72,8 +72,7 @@
     "lint:fix": "yarn lint --fix",
     "build:link": "node ./tasks/build-and-copy",
     "build:test-project": "node ./tasks/test-project/test-project",
-    "release": "node ./tasks/release/release.mjs",
-    "release-notes": "node ./tasks/release/release-notes.mjs",
+    "release": "node ./tasks/release/cli.mjs",
     "publish:canary": "lerna publish --force-publish --canary --include-merged-tags --preid canary --dist-tag canary --yes --loglevel verbose",
     "project:deps": "node ./tasks/framework-tools/frameworkDepsToProject.mjs",
     "project:copy": "node ./tasks/framework-tools/frameworkFilesToProject.mjs",
